Pull in improved code from test-observability repo
I think that when we created the current repo
(test-observability-action), we pulled in the code from the
test-observability-action-src repo, but we should have actually pulled
it from the test-observability repo, which contains the contents of
test-observability-action-src (see acb3b5f in test-observability), plus
a bunch of other changes, such as converting the action to TypeScript,
and adding GitHub actions for testing and linting.

These additional changes are all good things that we’d like to keep. So
in this commit, I’m copying across all of the action-related code from
test-observability at commit 4a75031.

Specifically, I’ve:

- Copied across actions/index.ts and deleted index.js

- Copied across actions/tsconfig.json, actions/.eslintrc.js, and .nvmrc

- Copied across dependency-related code from actions/package.json

- Manually added @types/node as a dev dependency – it seemed to be
  included as a transitive dependency in test-observability/action, not
  sure why that wasn’t the case here but haven’t spent too much time
  thinking about it

- Copied across the junit directory that contains test data

- Combined and copied across the action-related parts of the workflows
  action.yml and check.yml – I’ve skipped publish-action.yml which seems
  irrelevant to our current release process

I’ve opened ably/test-observability#67 to remove the unused action code
from the test-observability repo.

Co-authored-by: Morgan Adamiec <morgan.adamiec@ably.com>
Co-authored-by: Owen Pearson <owen.pearson@ably.com>