Upload a link to the GitHub job
Currently, the observability server UI links to the workflow run attempt
that the upload corresponds to. This is fine in the case where the
workflow only contains a single job, as it did in ably-cocoa. But in the
case where there’s more than one job in a workflow (e.g. in the case of
a matrix build), there’s no way to tell which job an upload corresponds
to. So, include this information in the upload.

I originally tried implementing this by passing a job-index input, whose
value was the index to which the current job corresponds in the response
from the "list jobs for a workflow run attempt" GitHub API. However,
some experimentation then showed that that the order in which the jobs
are listed in the action YAML file doesn’t necessarily match that in
which they are returned from the API, so there was no way to calculate
the value to pass for job-index. Hence, switched to using the job’s
`name` instead.

Resolves #18.